+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🐛 Bug Fix
 
 - Fix auto-completion of users having a localized display name.
+- Fix an issue with Twitch emotes & emojis in sent messages ([#3](https://github.com/HiDeoo/YaTA/issues/3)).
 
 # 0.8.0